Task0

+++ Tyuiu.LeushinP.Sprint3.Task0.V11.Lib/DataService.cs
@@ -5,16 +5,16 @@ namespace Tyuiu.LeushinP.Sprint2.Task0.V11.Lib
 {
     public class DataService : ISprint2Task0V11
     {
-        public bool[] GetCompareOperations(double x, double y)
+        public bool[] GetCompareOperations(int x, int y)
         {
             bool[] results = new bool[6];
 
-            results[0] = x == y;   // равно
-            results[1] = x != y;   // не равно
-            results[2] = x < y;    // меньше
-            results[3] = x > y;    // больше
-            results[4] = x <= y;   // меньше или равно
-            results[5] = x >= y;   // больше или равно
+            results[0] = x == y;
+            results[1] = x != y;
+            results[2] = x < y;   
+            results[3] = x > y;  
+            results[4] = x <= y;  
+            results[5] = x >= y;  
 
             return results;
         }
